fix: remove trail marks on line deletion #18
This change also deletes all trail marks on a given line or line range,
as this is a clear indication that they are no longer needed at the
specified position. Deleting only a word or the end of a line containing
trail marks will result in trail marks reappearing at the same position
as before deleting the word (if possible) or the end of the line on the
next motion.