fix: move session auto-loading into setup #30
This is necessary to account for Lazy.nvim's plugin load order, as files
within the `plugin` directory are sourced before the `setup` function is
called. This results in users not being able to enable the auto-load
feature as only the default values are present in the configuration
table at that time and auto-loading sessions is disabled by default.